kernel-hook: Rewrite main hook in Python
The primary motivation for this was to avoid using rsync pattern files
for specifying which device trees to copy. It was breaking for arm64,
because the files are organized by manufacturer, but I needed the .dtb
files themselves. There was also an issue with rsync pattern files
handling directories (or not handling them as I expected). All this, and
I was getting tired of writing complex shell script, and Python is
easier to read.